added the new landing page
